+++ packages/dart_http_client/lib/src/dart_http_client_transport.dart
@@ -110,6 +110,7 @@ final class DartHttpClientTransport implements HttpClientTransport {
       }
       httpRequest = bufferedRequest;
     }
+    httpRequest.followRedirects = request.redirectPolicy == DartHttpClientRedirectPolicy.follow;
     return httpRequest;
   }